All Done

+++ Project/Controller/Salesman.php
@@ -3,60 +3,109 @@
 class Controller_Salesman extends Controller_Admin_Action
 {
 	public function __construct()
-    {
-        if(!$this->authentication())
-        {
-            $this->redirect('login','admin_login');
-        }
-    }
-
-	public function gridAction()
 	{
-		$content = $this->getLayout()->getContent();
-		$salesmanGrid = Ccc::getBlock('Salesman_Grid');
-		$content->addChild($salesmanGrid,'grid');	
-		$this->setTitle('Salesman');
-		$this->renderLayout();
-	}
-
-	public function addAction()
-	{
-		$salesmanModel = Ccc::getModel('salesman');
-		$content = $this->getLayout()->getContent();
-		$this->setTitle('Salesman');
-		$salesmanAdd = Ccc::getBlock('Salesman_Edit')->setData(['salesman'=>$salesmanModel]);
-		$content->addChild($salesmanAdd,'add'); 
-		$this->renderLayout();
-	}
-
-	public function editAction()
-	{
-		try
+		if(!$this->authentication())
 		{
-			$this->setTitle('Salesman');
+			$this->redirect('login','admin_login');
+		}
+	}
+	public function indexAction()
+	{
+		$this->setTitle("Salesman");
+		$content = $this->getLayout()->getContent();
+		$salesmanIndex = Ccc::getBlock('Salesman_Index');
+		$content->addChild($salesmanIndex);
+		$this->renderLayout();
+	}
+
+	public function gridBlockAction()
+	{
+		
+		$salesmanGrid = Ccc::getBlock('Salesman_Grid')->toHtml();
+		$messageBlock = Ccc::getBlock('Core_Layout_Message')->toHtml();
+		$response = [
+			'status' => 'success',
+			'elements' => [
+				[
+					'element' => '#indexContent',
+					'content' => $salesmanGrid
+				],
+				[
+					'element' => '#adminMessage',
+					'content' => $messageBlock
+				]
+			]
+		];
+		$this->renderJson($response);
+	}
+	public function addBlockAction()
+	{
+		$salesmanModel = Ccc::getModel("Salesman");
+		$salesman = $salesmanModel;
+
+		Ccc::register('salesman',$salesman);
+
+		$salesmanEdit = Ccc::getBlock('Salesman_Edit')->toHtml();
+		$messageBlock = Ccc::getBlock('Core_Layout_Message')->toHtml();
+		$response = [
+			'status' => 'success',
+			'elements' => [
+				[
+					'element' => '#indexContent',
+					'content' => $salesmanEdit
+				],
+				[
+					'element' => '#salesmanMessage',
+					'content' => $messageBlock
+				]
+			]
+		];
+		$this->renderJson($response);
+		
+	}
+	public function editBlockAction()
+	{
+		try 
+   		{
+   			$salesmanModel = Ccc::getModel('Salesman');
 			$request = $this->getRequest();
-			$id = $request->getRequest('id');
-			if(!(int)$id)
+			$id = (int)$request->getRequest('id');
+			if(!$id)
 			{
-				throw new Exception("Invalid Request.", 1);	
+				throw new Exception("Invalid Request", 1);
 			}
-			$salesmenModel = Ccc::getModel('salesman');
-			$salesmen = $salesmenModel->load($id);
-			if(!$salesmen)
+			$salesman = $salesmanModel->load($id);
+			if(!$salesman)
 			{
-				throw new Exception("unable to fatch Record.", 1);
+				throw new Exception("System is unable to find record.", 1);
+				
 			}
-			$content = $this->getLayout()->getContent();
-			$salesmanEdit = Ccc::getBlock('Salesman_Edit')->setData(['salesman'=>$salesmen]);
-			$content->addChild($salesmanEdit,'edit'); 
-			$this->renderLayout();
-		}
-		catch(Exception $e)
-		{
-			$this->redirect('grid','salesman',[],true);
-		}
-	}
-	
+			$this->setTitle('Salesman Edit');
+			Ccc::register('salesman',$salesman);
+
+			$salesmanEdit = Ccc::getBlock('Salesman_Edit')->toHtml();
+			$messageBlock = Ccc::getBlock('Core_Layout_Message')->toHtml();
+			$response = [
+				'status' => 'success',
+				'elements' => [
+					[
+						'element' => '#indexContent',
+						'content' => $salesmanEdit
+					],
+					[
+						'element' => '#adminMessage',
+						'content' => $messageBlock
+					]
+				]
+			];
+			$this->renderJson($response);
+   		}	 
+   		catch (Exception $e) 
+   		{
+   			throw new Exception("Invalid Request.", 1);
+   		}
+   	}
+
 	public function saveAction()
 	{
 		try
@@ -89,11 +138,12 @@ class Controller_Salesman extends Controller_Admin_Action
 				
 			}
 			$this->getMessage()->addMessage('Salesman Inserted succesfully.',1); 
-			$this->redirect('grid','salesman',[],true);
+				$this->gridBlockAction();
+
 		}
 		catch(Exception $e)
 		{
-			$this->redirect('grid','salesman',[],true);	
+			$this->gridBlockAction();
 		}
 	}
 
@@ -106,43 +156,39 @@ class Controller_Salesman extends Controller_Admin_Action
 			$customerPriceModel = Ccc::getModel('Customer_Price');
 			$customerModel = Ccc::getModel('Customer');
 			$id = $request->getRequest('id');
-
 			if(!(int)$id)
 			{
 				throw new Exception("Invalid Request.", 1);
 				
 			}
-			/*print_r($id);
-			exit();*/
-			/*print_r("SELECT * FROM `customer` WHERE `salesmanId` = {$id}");
-			exit();*/
 			$customers = $customerModel->fetchAll("SELECT * FROM `customer` WHERE `salesmanId` = {$id}");
-
-			foreach($customers as $customer)
+			if($customers)
 			{
-				$customerPrices = $customerPriceModel->fetchAll("SELECT `entityId` FROM `customer_price` WHERE `customerId` = {$customer->customerId}");
-				
-				foreach ($customerPrices as $customerPrice) 
+				foreach($customers as $customer)
 				{
-					$customerPriceModel->load($customerPrice->entityId)->delete();
+					$customerPrices = $customerPriceModel->fetchAll("SELECT `entityId` FROM `customer_price` WHERE `customerId` = {$customer->customerId}");
+					if($customerPrices)
+					{
+
+					foreach($customerPrices as $customerPrice) 
+					{
+						$customerPriceModel->load($customerPrice->entityId)->delete();
+					}
+					}
 				}
 			}
-			
 			$delete = $salesmenModel->load($id)->delete();
-
-			
 			if(!$delete)
 			{
 				$this->getMessage()->addMessage('unable to delete.',3);
 				throw new Exception("unable to fatch Record.", 1);
 			}
-			
 			$this->getMessage()->addMessage('deleted succesfully.',1);
-			$this->redirect('grid','salesman',[],true);
+			$this->gridBlockAction();
 		}
 		catch(Exception $e)
 		{
-			$this->redirect('grid','salesman',[],true);
+			$this->gridBlockAction();
 		}
 	}
 }
